% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB.R
\name{ff}
\alias{ff}
\title{The Use of Marginal Distributions in Conditional Forecasting}
\usage{
ff(dt,m,w,n,q1)
}
\arguments{
\item{dt}{data frame}

\item{m}{the number of time series}

\item{w}{the number of predicted values}

\item{n}{gg}

\item{q1}{independent time series values}
}
\value{
the output from \code{ff()}
}
\description{
A new way to predict time series using the marginal distribution table in the absence of the significance of traditional models.
}
\examples{
x=rnorm(17,10,1)
y=rnorm(17,10,1)
data=data.frame(x,y)
print("Enter independent time series values")
q1=list(q=matrix(c(scan(,,quiet=TRUE)),1,2-1))
10.5


ff(data,2,1,17,q1)
}
